+++ HISTORY.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## x.x.x (not yet released)
+* Allow a Hint to be the first argument to Hint::build().
 
 ## 4.0.0 (2025-08-04)
 * Add `Hint::ONLY_FOR_DECODE` to support forward-compatibility of